# ModuleInstance.Functions/Exports as non-pointer slice 