--- a/internal/engine/interpreter/interpreter.go
+++ b/internal/engine/interpreter/interpreter.go
@@ -263,17 +263,17 @@ func (e *engine) CompileModule(ctx context.Context, module *wasm.Module, listene
 }
 
 // NewModuleEngine implements the same method as documented on wasm.Engine.
-func (e *engine) NewModuleEngine(name string, module *wasm.Module, importedFunctions []*wasm.FunctionInstance, moduleFunctions []wasm.FunctionInstance) (wasm.ModuleEngine, error) {
-	imported := uint32(len(importedFunctions))
+func (e *engine) NewModuleEngine(name string, module *wasm.Module, functions []wasm.FunctionInstance) (wasm.ModuleEngine, error) {
 	me := &moduleEngine{
-		name:                  name,
-		parentEngine:          e,
-		importedFunctionCount: imported,
+		name:         name,
+		parentEngine: e,
+		functions:    make([]*function, len(functions)),
 	}
 
-	for _, f := range importedFunctions {
+	imported := int(module.ImportFuncCount())
+	for i, f := range functions[:imported] {
 		cf := f.Module.Engine.(*moduleEngine).functions[f.Idx]
-		me.functions = append(me.functions, cf)
+		me.functions[i] = cf
 	}
 
 	codes, ok := e.getCodes(module)
@@ -282,9 +282,10 @@ func (e *engine) NewModuleEngine(name string, module *wasm.Module, importedFunct
 	}
 
 	for i, c := range codes {
-		f := &moduleFunctions[i]
+		offset := i + imported
+		f := &functions[offset]
 		inst := c.instantiate(f)
-		me.functions = append(me.functions, inst)
+		me.functions[offset] = inst
 	}
 	return me, nil
 }