--- a/internal/integration_test/bench/bench_test.go
+++ b/internal/integration_test/bench/bench_test.go
@@ -91,7 +91,7 @@ func BenchmarkCompilation(b *testing.B) {
 }
 
 func runCompilation(b *testing.B, r wazero.Runtime) wazero.CompiledModule {
-	compiled, err := r.CompileModule(testCtx, caseWasm)
+	compiled, err := r.CompileModule(testCtx, inita)
 	if err != nil {
 		b.Fatal(err)
 	}