… internal/integration_test/bench/hostfunc_bench_test.go
… internal/wasm/module.go
… internal/wasm/module_test.go
… internal/wasm/store_test.go
… internal/wasm/table_test.go
… runtime_test.go