Ref #889 

Here's a benchmark result for `BenchmarkInitialization` with [libcre2.so](https://github.com/anuraaga/re2-go/blob/main/wasm/libcre2.so) binary by @anuraaga 

```
$ benchstat old_re2.txt new_re2.txt
name                                    old time/op    new time/op    delta
Initialization/interpreter-10              152µs ± 3%     139µs ± 2%   -8.39%  (p=0.008 n=5+5)
Initialization/interpreter-multiple-10    62.4µs ± 2%    74.4µs ±38%     ~     (p=0.730 n=4+5)
Initialization/compiler-10                 197µs ±24%     170µs ± 4%  -13.51%  (p=0.008 n=5+5)
Initialization/compiler-multiple-10       58.2µs ±16%    44.2µs ±14%  -24.01%  (p=0.016 n=5+5)

name                                    old alloc/op   new alloc/op   delta
Initialization/interpreter-10              630kB ± 0%     553kB ± 0%  -12.28%  (p=0.008 n=5+5)
Initialization/interpreter-multiple-10     631kB ± 0%     554kB ± 0%  -12.25%  (p=0.008 n=5+5)
Initialization/compiler-10                 543kB ± 0%     509kB ± 0%   -6.21%  (p=0.008 n=5+5)
Initialization/compiler-multiple-10        544kB ± 0%     510kB ± 0%   -6.21%  (p=0.008 n=5+5)

name                                    old allocs/op  new allocs/op  delta
Initialization/interpreter-10              1.92k ± 0%     1.88k ± 0%   -1.83%  (p=0.008 n=5+5)
Initialization/interpreter-multiple-10     1.92k ± 0%     1.89k ± 0%   -1.82%  (p=0.000 n=5+4)
Initialization/compiler-10                  62.0 ± 0%      40.0 ± 0%  -35.48%  (p=0.008 n=5+5)
Initialization/compiler-multiple-10         69.0 ± 0%      47.0 ± 0%  -31.88%  (p=0.008 n=5+5)

```
### `old_re2.txt`

```
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
BenchmarkInitialization/interpreter-10         	    7048	    149845 ns/op	  630451 B/op	    1915 allocs/op
BenchmarkInitialization/interpreter-multiple-10         	   16717	     61864 ns/op	  631379 B/op	    1922 allocs/op
BenchmarkInitialization/compiler-10                     	    7623	    244489 ns/op	  542738 B/op	      62 allocs/op
BenchmarkInitialization/compiler-multiple-10            	   28168	     65375 ns/op	  543597 B/op	      69 allocs/op
PASS
ok  	github.com/tetratelabs/wazero/internal/integration_test/bench	12.556s
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
BenchmarkInitialization/interpreter-10         	    6861	    149691 ns/op	  630450 B/op	    1915 allocs/op
BenchmarkInitialization/interpreter-multiple-10         	   16336	     63583 ns/op	  631306 B/op	    1922 allocs/op
BenchmarkInitialization/compiler-10                     	    6507	    179949 ns/op	  542747 B/op	      62 allocs/op
BenchmarkInitialization/compiler-multiple-10            	   27397	     58223 ns/op	  543572 B/op	      69 allocs/op
PASS
ok  	github.com/tetratelabs/wazero/internal/integration_test/bench	12.309s
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
BenchmarkInitialization/interpreter-10         	    6944	    149922 ns/op	  630450 B/op	    1915 allocs/op
BenchmarkInitialization/interpreter-multiple-10         	   18090	     61727 ns/op	  631365 B/op	    1922 allocs/op
BenchmarkInitialization/compiler-10                     	    8559	    181302 ns/op	  542739 B/op	      62 allocs/op
BenchmarkInitialization/compiler-multiple-10            	   28434	     59833 ns/op	  543567 B/op	      69 allocs/op
PASS
ok  	github.com/tetratelabs/wazero/internal/integration_test/bench	11.860s
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
BenchmarkInitialization/interpreter-10         	    6868	    156870 ns/op	  630453 B/op	    1915 allocs/op
BenchmarkInitialization/interpreter-multiple-10         	   17752	     62528 ns/op	  631370 B/op	    1922 allocs/op
BenchmarkInitialization/compiler-10                     	    6874	    191044 ns/op	  542742 B/op	      62 allocs/op
BenchmarkInitialization/compiler-multiple-10            	   30200	     58636 ns/op	  543610 B/op	      69 allocs/op
PASS
ok  	github.com/tetratelabs/wazero/internal/integration_test/bench	11.311s
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
BenchmarkInitialization/interpreter-10         	    6967	    151888 ns/op	  630452 B/op	    1915 allocs/op
BenchmarkInitialization/interpreter-multiple-10         	   15666	     83863 ns/op	  631391 B/op	    1922 allocs/op
BenchmarkInitialization/compiler-10                     	    5769	    186195 ns/op	  542755 B/op	      62 allocs/op
BenchmarkInitialization/compiler-multiple-10            	   23019	     48930 ns/op	  543521 B/op	      69 allocs/op
PASS
ok  	github.com/tetratelabs/wazero/internal/integration_test/bench	9.294s

```


### `new_re2.txt`

```
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
BenchmarkInitialization/interpreter-10         	   43489	     25071 ns/op	  141474 B/op	      85 allocs/op
BenchmarkInitialization/interpreter-multiple-10         	  103897	     11115 ns/op	  142588 B/op	      92 allocs/op
BenchmarkInitialization/compiler-10                     	   36500	     33019 ns/op	  144560 B/op	      35 allocs/op
BenchmarkInitialization/compiler-multiple-10            	  101678	     18000 ns/op	  145272 B/op	      41 allocs/op
PASS
ok  	github.com/tetratelabs/wazero/internal/integration_test/bench	6.615s
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
BenchmarkInitialization/interpreter-10         	   51102	     31718 ns/op	  141474 B/op	      85 allocs/op
BenchmarkInitialization/interpreter-multiple-10         	   91940	     13639 ns/op	  142592 B/op	      92 allocs/op
BenchmarkInitialization/compiler-10                     	   30066	     36717 ns/op	  144560 B/op	      35 allocs/op
BenchmarkInitialization/compiler-multiple-10            	   97819	     15121 ns/op	  145301 B/op	      41 allocs/op
PASS
ok  	github.com/tetratelabs/wazero/internal/integration_test/bench	6.837s
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
BenchmarkInitialization/interpreter-10         	   53106	     21618 ns/op	  141473 B/op	      85 allocs/op
BenchmarkInitialization/interpreter-multiple-10         	  108909	     10744 ns/op	  142691 B/op	      92 allocs/op
BenchmarkInitialization/compiler-10                     	   36554	     33729 ns/op	  144559 B/op	      35 allocs/op
BenchmarkInitialization/compiler-multiple-10            	  103538	     17683 ns/op	  145264 B/op	      41 allocs/op
PASS
ok  	github.com/tetratelabs/wazero/internal/integration_test/bench	6.637s
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
BenchmarkInitialization/interpreter-10         	   54764	     21388 ns/op	  141473 B/op	      85 allocs/op
BenchmarkInitialization/interpreter-multiple-10         	  107838	     10712 ns/op	  142695 B/op	      92 allocs/op
BenchmarkInitialization/compiler-10                     	   35791	     32797 ns/op	  144560 B/op	      35 allocs/op
BenchmarkInitialization/compiler-multiple-10            	   98738	     16353 ns/op	  145272 B/op	      41 allocs/op
PASS
ok  	github.com/tetratelabs/wazero/internal/integration_test/bench	6.384s
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
BenchmarkInitialization/interpreter-10         	   54931	     22534 ns/op	  141473 B/op	      85 allocs/op
BenchmarkInitialization/interpreter-multiple-10         	  107678	     10865 ns/op	  142697 B/op	      92 allocs/op
BenchmarkInitialization/compiler-10                     	   36722	     32671 ns/op	  144559 B/op	      35 allocs/op
BenchmarkInitialization/compiler-multiple-10            	   91105	     14615 ns/op	  145270 B/op	      41 allocs/op
PASS
ok  	github.com/tetratelabs/wazero/internal/integration_test/bench	7.212s
```